Tables dictate threads

+++ brimming_test.go
@@ -18,7 +18,7 @@ func TestStrToFloat(t *testing.T) {
 }
 
 func TestSizeToRows(t *testing.T) {
-	expect := map[string]int{
+	expect := map[string]int64{
 		"2 thing": 0,
 		"100mb":   100000,
 		"123gb":   123000000,
